move replacement routines and common labels into FML folder. Fix name of FML
