+++ gusto/fml/__init__.py
@@ -1,1 +1,2 @@
-from gusto.fml.form_manipulation_labelling import *        # noqa
+from gusto.fml.form_manipulation_language import *        # noqa
+from gusto.fml.replacement import *                       # noqa
